requirements.txt
alembic==0.9.9
bcrypt==3.1.4
beautifulsoup4==4.4.1
blinker==1.4
Brlapi==0.6.4
cffi==1.11.5
chardet==2.3.0
click==6.7
command-not-found==0.3
defer==1.0.6
dominate==2.3.1
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.2.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
gunicorn==19.9.0
html5lib==0.999
itsdangerous==0.24
Jinja2==2.10
language-selector==0.1
louis==2.6.4
lxml==3.5.0
Mako==1.0.7
MarkupSafe==1.0
onboard==1.2.0
pexpect==4.0.1
Pillow==3.1.2
psycopg2==2.7.5
psycopg2-binary==2.7.5
ptyprocess==0.5
pycparser==2.18
pycups==1.9.73
pycurl==7.43.0
pygobject==3.20.0
python-apt===1.1.0.b1-elementary2-r2-ubuntu0.4.1.1
python-dateutil==2.7.3
python-editor==1.0.3
python-systemd==231
pyxdg==0.25
reportlab==3.3.0
requests==2.9.1
sessioninstaller==0.0.0
six==1.11.0
SQLAlchemy==1.2.8
system-service==0.3
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
urllib3==1.13.1
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.2.1
xkit==0.0.0
